refactor to shared factories and shared tests
