+++ packages/tsrx-solid/tests/basic.test.js
@@ -1,5 +1,12 @@
 import { describe, expect, it } from 'vitest';
+import { runSharedSourceMappingTests } from '@tsrx/core/test-harness/source-mappings';
 import { compile, compile_to_volar_mappings } from '../src/index.js';
+
+runSharedSourceMappingTests({
+	compile_to_volar_mappings,
+	name: 'solid',
+	rejectsComponentAwait: true,
+});
 
 describe('@tsrx/solid basic', () => {
 	describe('component → function', () => {
