+++ packages/tsrx/package.json
@@ -26,7 +26,8 @@
     },
     "./types/acorn": {
       "types": "./types/acorn.d.ts"
-    }
+    },
+    "./test-harness/source-mappings": "./tests/shared/source-mappings.js"
   },
   "dependencies": {
     "@jridgewell/sourcemap-codec": "catalog:default",
